align folder structure

+++ features/movieList/src/main/java/com/example/movielist/domain/model/MovieDomain.kt
@@ -1,8 +1,8 @@
 package com.example.movielist.domain.model
 
 
-data class MovieDomain(
-    val id: Int,
+data class  MovieDomain(
+    val id: Long,
     val title: String,
     val image: String,
     val year: String,
